Making va-crisis-line-modal a standalone component
