+++ packages/web-components/src/components/va-minimal-header/va-minimal-header.tsx
@@ -1,6 +1,12 @@
 import { Component, Host, h, Prop} from '@stencil/core';
 import vaSeal from '../../assets/va-seal.svg';
 
+/**
+ * @componentName Minimal Header
+ * @maturityCategory caution
+ * @maturityLevel candidate
+ */
+
 @Component({
   tag: 'va-minimal-header',
   styleUrl: 'va-minimal-header.scss',
@@ -14,13 +20,13 @@ export class VaMinimalHeader {
 
   render() {
     const {header, subheader} = this;
-    
+
     return (
       <Host>
         <va-official-gov-banner/>
         <va-crisis-line-modal/>
         <div class="va-header">
-          
+
           <a href="/" >
             <img class="va-logo" src={vaSeal} aria-hidden="true"/>
           </a>
